Refactored into models and resources.
