+++ .gitignore
@@ -128,3 +128,5 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
+# SQLite databases
+*.db
