+++ security.py
@@ -1,13 +1,13 @@
 from werkzeug.security import safe_str_cmp
-from user import User
+from models.user import UserModel
 
 
 def authenticate(username, password):
-    user = User.find_by_username(username)
+    user = UserModel.find_by_username(username)
     if user and safe_str_cmp(user.password, password):
         return user
 
 
 def identity(payload):
     user_id = payload['identity']
-    return User.find_by_id(user_id)
+    return UserModel.find_by_id(user_id)
